Made a few changes to allow expansion for capabilities and resolves some reviews by team members
