+++ src/main/java/org/kainos/ea/controller/JobRoleController.java
@@ -2,6 +2,7 @@ package org.kainos.ea.controller;
 
 import io.swagger.annotations.Api;
 import org.kainos.ea.dao.JobRoleDao;
+import org.kainos.ea.exception.DatabaseConnectionException;
 import org.kainos.ea.service.JobRoleService;
 import org.kainos.ea.exception.FailedToGetJobRoleException;
 
@@ -26,20 +27,26 @@ public class JobRoleController {
             return Response.ok(jobRoleService.getAllJobRoles()).build();
         } catch (FailedToGetJobRoleException e) {
             System.err.println((e.getMessage()));
+            return Response.status(Response.Status.BAD_REQUEST).build();
+        } catch (DatabaseConnectionException e) {
+            System.err.println((e.getMessage()));
             return Response.serverError().build();
         }
     }
 
-//    @GET
-//    @Path("/job-roles/{id}")
-//    @Produces(MediaType.APPLICATION_JSON)
-//    public Response getJobRoleById(@PathParam("id") int id) {
-//        try {
-//            return Response.ok(jobRoleService.getJobRolesById(id)).build();
-//        } catch (FailedToGetJobRoleException e) {
-//            System.err.println((e.getMessage()));
-//            return Response.serverError().build();
-//        }
-//    }
+    @GET
+    @Path("/job-roles/{id}")
+    @Produces(MediaType.APPLICATION_JSON)
+    public Response getJobRoleById(@PathParam("id") int id) {
+        try {
+            return Response.ok(jobRoleService.getJobRolesById(id)).build();
+        } catch (DatabaseConnectionException e) {
+            System.err.println((e.getMessage()));
+            return Response.serverError().build();
+        } catch (FailedToGetJobRoleException e) {
+            System.err.println((e.getMessage()));
+            return Response.status(Response.Status.BAD_REQUEST).build();
+        }
+    }
 
 }
