+++ src/main/java/org/kainos/ea/dao/JobRoleDao.java
@@ -1,5 +1,6 @@
 package org.kainos.ea.dao;
 
+import org.kainos.ea.exception.DatabaseConnectionException;
 import org.kainos.ea.model.Band;
 import org.kainos.ea.model.Capability;
 import org.kainos.ea.model.JobRole;
@@ -11,7 +12,7 @@ import java.util.List;
 public class JobRoleDao {
 
 
-    public List<JobRole> getAllJobRoles() throws SQLException {
+    public List<JobRole> getAllJobRoles() throws SQLException, DatabaseConnectionException {
         Connection c = DatabaseConnector.getConnection();
 
         assert c != null;
@@ -45,23 +46,29 @@ public class JobRoleDao {
         return jobRoleList;
     }
 
-//    public JobRole getJobRoleById(int id) throws SQLException{
-//        Connection c = DatabaseConnector.getConnection();
-//
-//        Statement st = c.createStatement();
-//
-//        ResultSet rs = st.executeQuery("SELECT id, name, band_id, specification FROM job_roles WHERE id = " + id + ";");
-//
-//        while (rs.next()) {
-//            return new JobRole(
-//                    rs.getInt("id"),
-//                    rs.getString("name"),
-//                    rs.getInt("band_id"),
-//                    rs.getString("specification")
-//            );
-//
-//        }
-//
-//        return null;
-//    }
+    public JobRole getJobRoleById(int id) throws SQLException, DatabaseConnectionException {
+        Connection c = DatabaseConnector.getConnection();
+
+        Statement st = c.createStatement();
+
+        ResultSet rs = st.executeQuery("SELECT job_roles.id, job_roles.name, bands.`name`, level, job_roles.specification  FROM job_roles JOIN bands ON job_roles.band_id = bands.id " +
+                "JOIN capabilities ON job_roles.capability_id = capabilities.id WHERE job_roles.id = " + id + ";");
+
+        while (rs.next()) {
+
+            Band band = new Band();
+            band.setName(rs.getString("bands.name"));
+            band.setLevel(rs.getInt("level"));
+
+            return new JobRole(
+                    rs.getInt("job_roles.id"),
+                    rs.getString("job_roles.name"),
+                    band,
+                    rs.getString("job_roles.specification")
+            );
+
+        }
+
+        return null;
+    }
 }
