+++ src/main/java/org/kainos/ea/exception/DatabaseConnectionException.java
@@ -1,4 +1,4 @@
-package org.kainos.ea.client;
+package org.kainos.ea.exception;
 
 public class DatabaseConnectionException extends Throwable {
